New button for layouting edges only
The KIELER KLay layouters do not support layouting of edges only. There
is a C++ library called libavoid that supports routing of the edges in
a diagram with fixed node positions. It includes a build script for
generating JNI interface classes.

KIELER has a LayoutProvider implementation for libavoid
(LibavoidLayoutProvider) so that we can use the KGraph data model and
the same model transformation logic as for the normal layouter
(layered layouter).

libavoid is added as a submodule to track the version of the binaries in
the runtime/lib/libavoid folder. I don't use the latest libavoid
version, because there are compile errors. The referenced version is
latest that I was able to build.

This is the first working version. There is some fine tuning required
(if possible). At the moment there is only a native library of libavoid
for OSX. I haven't compiled it for Windows and Linux yet.

I plan to use libavoid to reroute edges while interactively moving nodes
around. That's the more useful use case. The current global edge routing
is more like the first step to that instead of the reason for adding
libavoid.