Routing of selected edges only
If one or more edges are selected and the 'Layout Edges' button is
pressed, only the selected edges are rerouted. The routes of the other
edges are fixed but they are considered by the layouter.
If nothing is selected all edges are rerouted.

The is still a bug that causes not selected edges being changed.