The example language for activity diagrams
A StatementList is shown as a diagram. It has start/end states and implicit connection defined by the order of the statements.

Showing boxes without defining a the special "box" editor for them theme to work properly now. All nodes returned in the content section that do not have a diagram specific editor are shown as boxes.