PortLayouter performance
The ports were move just far enough to not overlap with its neighbour, but the next neighbour pushed it back.
After some iterations, the overlap was very small and the move distance also, getting smaller and smaller.
Now the ports are moved a bit more apart than it is required to make them not overlap.
This makes the algorithm reach the exit condition (no overlaps) faster.

Before the timeout of 1000 iterations was always reached. Now it exits after about 10 iterations.