Reduced number of repaints on mouse move
The diagram was repainted on each mouse move.
This was necessary when the port label opacity decreased with the distance of the mouse cursor.
Now the opacity is toggled between 0% and 100% depending on the mouse cursor being inside or outside a box.

The diagram is only redrawn, if the opacity of any label changed.