Smart guides are not allowed to cross boxes
When there are a lot of boxes in the diagram, there are a lot of snap points which makes them useless.
Therefore, we only use smart guide of the boxes directly next to the moved box.
If a box is 'hidden' behind another box (the guide would cross this other box), its smart guides do not apply.