Fixed wrapping in embedded MPS cells
The content of a box were wrapped depending on its position.
If it was close to or on the right side of the vertical "wrapper" line, the content was wrapped, but in a box more on the left no wrapping happened.
Usually, you don't want any wrapping of long lines inside a box.
Fixed it by moving a cell to (0;0) before calling relayout.

See #152