After a delete the selection was set to a random position
It was not random, but always set to an unexpected position. The window
then scrolls to the selected cell, which is annoying if it is a totally
wrong position.
The default selection handling from MPS doesn't make sense here.

Now the selection is set the diagram and the windows never scrolls when
the diagram receives the selection.

#160